refactor(UsageAnalytics): modularize usage analytics page
Move the usage analytics page into a feature module to improve
maintainability and reusability. This includes:

- Moving the main page to `features/UsageAnalytics/UsageAnalyticsPage.tsx`
- Extracting chart helpers to `features/UsageAnalytics/charts/echartsOptions.ts`
- Extracting types to `features/UsageAnalytics/types.ts`
- Extracting hooks to `features/UsageAnalytics/hooks/`
- Extracting components to `features/UsageAnalytics/components/`
- Creating a barrel file for the feature module
