+++ features/UsageAnalytics/hooks/useUsageAnalyticsExport.ts
@@ -1,0 +1,52 @@
+import { useCallback } from "react"
+import toast from "react-hot-toast"
+import { useTranslation } from "react-i18next"
+
+import { computeUsageHistoryExport } from "~/services/history/usageHistory/analytics"
+import type {
+  UsageHistoryExportSelection,
+  UsageHistoryStore,
+} from "~/types/usageHistory"
+import { getErrorMessage } from "~/utils/error"
+
+export const useUsageAnalyticsExport = (params: {
+  store: UsageHistoryStore | null
+  exportSelection: UsageHistoryExportSelection | null
+}) => {
+  const { store, exportSelection } = params
+  const { t } = useTranslation("usageAnalytics")
+
+  const handleExport = useCallback(async () => {
+    if (!store || !exportSelection) {
+      toast.error(t("messages.error.exportNoData"))
+      return
+    }
+
+    try {
+      const exportData = computeUsageHistoryExport({
+        store,
+        selection: exportSelection,
+      })
+      const blob = new Blob([JSON.stringify(exportData, null, 2)], {
+        type: "application/json",
+      })
+      const url = URL.createObjectURL(blob)
+      const link = document.createElement("a")
+      link.href = url
+      link.download = `all-api-hub-usage-history-${
+        new Date().toISOString().split("T")[0]
+      }.json`
+      document.body.appendChild(link)
+      link.click()
+      document.body.removeChild(link)
+      URL.revokeObjectURL(url)
+      toast.success(t("messages.success.exported"))
+    } catch (error) {
+      toast.error(
+        t("messages.error.exportFailed", { error: getErrorMessage(error) }),
+      )
+    }
+  }, [exportSelection, store, t])
+
+  return { handleExport }
+}
