+++ features/UsageAnalytics/types.ts
@@ -1,0 +1,8 @@
+export type UsageAnalyticsChartDisplayType = "pie" | "bar"
+
+export type UsageAnalyticsBreakdownChartKey =
+  | "slowModels"
+  | "slowTokens"
+  | "accountComparison"
+  | "modelDistribution"
+  | "modelCostDistribution"
